requirements-ml.txt
torch==2.8.0
transformers==4.40.0
tokenizers==0.19.0
scikit-learn==1.4.0

pandas==2.2.0
numpy==1.26.3

notebook==7.0.7
jupyterlab==4.0.11

matplotlib==3.8.2
seaborn==0.12.2

psutil==5.9.5

streamlit==1.26.1
Pillow==9.5.0
python-dotenv==1.0.0

langdetect==1.0.9
googletrans==4.0.0rc1

imagehash==4.3.1
ExifRead==3.0.0
scipy==1.10.1
opencv-python==4.8.0
opencv-python-headless==4.8.0

moviepy==1.0.3

tqdm==4.66.1
scikit-image==0.22.3
